frontend/requirements.txt
Pillow==11.0.0
Requests==2.32.3
streamlit==1.38.0
